feature: implemented auth service and token handling

+++ Models/Producto.cs
@@ -6,7 +6,7 @@
         {
         }
 
-        public Producto(string descripcion, int stock, double precio, Categoria categoria)
+        public Producto(string descripcion, int stock, decimal precio, Categoria categoria)
         {
             Descripcion = descripcion;
             Stock = stock;
@@ -17,7 +17,7 @@
         public int Id { get; private set; }
         public string Descripcion { get; private set; }
         public int Stock { get; private set; }
-        public double Precio { get; private set; }
+        public decimal Precio { get; private set; }
         public Categoria Categoria { get; private set; }
     }
